# shared/openapi/openapi.json
{
  "components" : {
    "schemas" : {
...
      "CollectionModelPermissionResponse" : {
        "properties" : {
          "_embedded" : {
            "properties" : {
              "PermissionCollection" : {
                "items" : {
                  "$ref" : "#/components/schemas/Permission"
                },
                "type" : "array"
              }
            },
            "type" : "object"
          },
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          }
        },
        "type" : "object"
      },
      "CollectionModelRoleResponse" : {
        "properties" : {
          "_embedded" : {
            "properties" : {
              "RoleCollection" : {
                "items" : {
                  "$ref" : "#/components/schemas/Role"
                },
                "type" : "array"
              }
            },
            "type" : "object"
          },
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          }
        },
        "type" : "object"
      },
      "CollectionModelUserDetailsResponse" : {
        "properties" : {
          "_embedded" : {
            "properties" : {
              "UserDetailsCollection" : {
                "items" : {
                  "$ref" : "#/components/schemas/UserDetails"
                },
                "type" : "array"
              }
            },
            "type" : "object"
          },
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          }
        },
        "type" : "object"
      },
      "CollectionModelUserInfoResponse" : {
        "properties" : {
          "_embedded" : {
            "properties" : {
              "UserInfoCollection" : {
                "items" : {
                  "$ref" : "#/components/schemas/UserInfo"
                },
                "type" : "array"
              }
            },
            "type" : "object"
          },
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          }
        },
        "type" : "object"
      },
...
      "Link" : {
        "properties" : {
          "deprecation" : {
            "type" : "string"
          },
          "href" : {
            "type" : "string"
          },
          "hreflang" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "profile" : {
            "type" : "string"
          },
          "templated" : {
            "type" : "boolean"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "Links" : {
        "additionalProperties" : {
          "$ref" : "#/components/schemas/Link"
        },
        "type" : "object"
      },
      "Permission" : {
        "properties" : {
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          },
          "description" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "RequestError" : {
        "properties" : {
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          },
          "details" : {
            "type" : "object"
          },
          "error" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "Role" : {
        "properties" : {
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          },
          "description" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "permissions" : {
            "items" : {
              "$ref" : "#/components/schemas/Permission"
            },
            "type" : "array",
            "uniqueItems" : true
          }
        },
        "type" : "object"
      },
      "RoleForm" : {
        "properties" : {
          "description" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        },
        "required" : [ "name" ],
        "type" : "object"
      },
...
      "UserDetails" : {
        "properties" : {
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          },
          "email" : {
            "type" : "string"
          },
          "roles" : {
            "items" : {
              "$ref" : "#/components/schemas/Role"
            },
            "type" : "array",
            "uniqueItems" : true
          },
          "username" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "UserInfo" : {
        "properties" : {
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          },
          "points" : {
            "format" : "int32",
            "type" : "integer"
          },
          "username" : {
            "type" : "string"
          }
        },
        "type" : "object"
      }
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.0.1",
  "paths" : {
    "/administration/permissions" : {
      "get" : {
        "operationId" : "getPermissions",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CollectionModelPermissionResponse"
                }
              }
            },
            "description" : "List of all permissions"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Bad request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "tags" : [ "administration-controller" ]
      }
    },
    "/administration/roles" : {
      "get" : {
        "operationId" : "getRoles",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CollectionModelRoleResponse"
                }
              }
            },
            "description" : "List of all roles"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Bad request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "tags" : [ "administration-controller" ]
      }
    },
    "/administration/roles/{username}" : {
      "delete" : {
        "operationId" : "removeRoleFromUser",
        "parameters" : [ {
          "in" : "path",
          "name" : "username",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RoleForm"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserDetails"
                }
              }
            },
            "description" : "Add role to user"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Bad request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "tags" : [ "administration-controller" ]
      },
      "put" : {
        "operationId" : "addRoleToUser",
        "parameters" : [ {
          "in" : "path",
          "name" : "username",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RoleForm"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserDetails"
                }
              }
            },
            "description" : "Add role to user"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Bad request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "tags" : [ "administration-controller" ]
      }
    },
    "/administration/users" : {
      "get" : {
        "operationId" : "getUsers_1",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CollectionModelUserDetailsResponse"
                }
              }
            },
            "description" : "List of all users"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Bad request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "tags" : [ "administration-controller" ]
      }
    },
...
    "/users/" : {
      "get" : {
        "operationId" : "getUsers",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CollectionModelUserInfoResponse"
                }
              }
            },
            "description" : "User is authenticated"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Bad request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestError"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "tags" : [ "user-controller" ]
      }
    }
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}